+++ experimental/packages/configuration/test/FileConfigFactory.test.ts
@@ -194,7 +194,6 @@ const configFromKitchenSinkFile = {
       },
       detectors: [
         { container: null },
-        { env: null },
         { host: null },
         { os: null },
         { process: null },