+++ packages/opentelemetry-resources/src/detectors/platform/browser/ServiceDetector.ts
@@ -5,7 +5,4 @@
 
 import { noopDetector } from '../../NoopDetector';
 
-/**
- * @experimental
- */
-export const serviceInstanceIdDetector = noopDetector;
+export const serviceDetector = noopDetector;